Updates Python module loading process
Modifies maali such that it does not load a Python interpreter when one is not already loaded, but throws an error instead. This is
avoid users relying on a default setting that may change in the future. The user must explicitly load a proper Python interpreter before
loading a Python module.